Inventory improvements, Bug fix
Fixed an issue I was having with Editors not working, drastically improved inventory layout. Also did some refactoring of the less layout to make things easier on me.

Added a placeholder "Details" tab to items, which I will be setting up soon enough (it was necessary to get the tab navigation working right on item sheets)